Reorganized structure of the package.
