--- modules/generator.py
@@ -1,1 +1,0 @@
-def generator() -> None: ...
