+++ testcase/modules/compare.py
@@ -1,0 +1,1 @@
+def compare() -> None: ...
