+++ testcase/modules/generate.py
@@ -1,0 +1,1 @@
+def generate() -> None: ...
