+++ testcase/tests/__init__.py
@@ -1,0 +1,4 @@
+from .compare_test import TestCompare
+from .compile_test import TestCompile
+from .generator_test import TestGenerator
+from .init_test import TestInit
