+++ testcase/tests/compare_test.py
@@ -1,0 +1,4 @@
+import unittest
+
+
+class TestCompare(unittest.TestCase): ...
