+++ testcase/tests/compile_test.py
@@ -1,0 +1,4 @@
+import unittest
+
+
+class TestCompile(unittest.TestCase): ...
