+++ testcase/tests/generator_test.py
@@ -1,0 +1,4 @@
+import unittest
+
+
+class TestGenerator(unittest.TestCase): ...
