+++ testcase/tests/init_test.py
@@ -1,0 +1,4 @@
+import unittest
+
+
+class TestInit(unittest.TestCase): ...
