+++ testcase/utils/compile/cpp.py
@@ -1,0 +1,19 @@
+from utils.compile.language import Language
+
+
+class Cpp(Language):
+    def __init__(self) -> None:
+        standards = ["c++98", "c++03", "c++11", "c++14", "c++17"]
+        super().__init__(standards)
+
+    def compileToExecutive(self, source: str, std: str, args: list) -> str:
+        """Compiles the cpp source code to executive.
+
+        Args:
+            source: (str) The absolute route of the source.
+            std: (str) The name of language standard.
+            args: (list) The list of arguments for the compiler or interpreter.
+        Returns:
+            The absolute route of executive.
+        """
+        raise NotImplementedError()
