+++ testcase/utils/compile/rust.py
@@ -1,0 +1,6 @@
+from utils.compile.language import Language
+
+
+class Rust(Language):
+    def __init__(self) -> None:
+        raise NotImplementedError()
