+++ unit_test.py
@@ -2,7 +2,7 @@
 
 import unittest
 
-from tests import *
+from testcase.tests import *
 
 if __name__ == "__main__":
     unittest.main(verbosity=2)
